Separate dashboard functions by application
All functions to create one panel on the dashboard have their own file now
to allow finding them later more easily.
